Move release plan milestone to legacy file
